containers/zmq_consumer/requirements.txt
pyzmq
requests
python-bitcoinrpc
tqdm
pprint
pandas
simplejson